On the way to the next release: Bugfixes and the beginning of gradient filling.
svn path=/trunk/koffice/; revision=11795

+++ killustrator/xmlutils/XmlReader.cc
@@ -211,7 +211,7 @@ bool XmlReader::readAttributes (list<XmlAttribute>& attrib_list) {
       break;
     }
     else {
-      cout << "invalid attribute name: " << tok << endl;
+      cout << "invalid attribute name: " << (int) tok << endl;
       return false;
     }
     tok = tokenizer.nextToken ();
